+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-04 20:24:54
-// Commit: test(core): implement widget integration tests for core
+// Updated at: 2026-05-05 13:32:33
+// Commit: test(core): add unit tests for core business logic
 
 class coreComponentState {
-  final String commitRef = '855';
-  final String lastMessage = 'test(core): implement widget integration tests for core';
-  final DateTime timestamp = DateTime.parse('2026-05-04T20:24:54+05:00');
+  final String commitRef = '882';
+  final String lastMessage = 'test(core): add unit tests for core business logic';
+  final DateTime timestamp = DateTime.parse('2026-05-05T13:32:33+05:00');
 }